Don't sort when listing names
The documentation says that the order of names being visited with
listFullyScopedNames and listLocallyScopedNames is unspecified.

The previous commit provided support for listing in the order the items
were added to the configuration.

However, it seems that the default implementation of
listFullyScopedNames and listLocallyScopedNames sort the array of names
before returning.

Technically, this qualifies as unspecified, but it means that the
original order is lost when sorted.