Fix search for override and fallback configs
Here is the signature for ConfigurationImpl::lookup.

    ConfigItem *
    ConfigurationImpl::lookup(
        const char * fullyScopedName,
        const char * localName,
        bool startInRoot,
        bool searchOutwards) const;

However, no matter what the name says, fullyScopedName is not always
fully scoped.  Instead, it is the scoped name relative to the lookup
scope, which is not always the root scope.

If fullyScopedName starts with '.' then this is an explicit absolute
name, which will always be searched relative to the root scope.

If startInRoot is true, then the search will be performed relative to
the root scope.

Otherwise, the search will be performed relative to the current
scope, which may not be the root scope.  In particular, while
performing a parse, the current scope will be whatever explicit
nesting level is currently being parsed.  Specifically, when parsing
'a.b.c = "foo";' the current scope is the root scope, but when
parsing 'a { b { c = "foo"; } }' the current scope is 'a.b'.

In such a case, fullyScopedName will simply be the string "c" even
though the fully scoped name, relative to the root scope, is "a.b.c"

However, the current scope for the override and fallback
configurations will be the root scope in all such cases.  Thus, when
delegating the search to one of them, we need to make sure we start
the search at the root, which also means that we need to create a
true fully scoped name, relative to the root.